refactor(simulation): decompose weaponAttackHitResolution.ts into 2 modules
The single `resolveWeaponHit` function had grown to 548 lines, exceeding
the oxlint `max-lines: 400` rule. Extracted six cohesive sub-steps into a
co-located `weaponAttackHitResolution.helpers.ts` sibling:

- `emitCoveredLegMiss` — Total Warfare p.53 partial-cover leg-hit miss
- `consumeWeaponAmmo` — ammo bin decrement + `AmmoConsumed` emission
- `emitDamageChainEvents` — `DamageApplied`/`LocationDestroyed`/
  `TransferDamage` walk incl. side-torso -> arm cascade (largest block)
- `emitHeadHitPilotEvent` — raw head-hit `PilotHit` with crit de-dupe
- `emitUnitDestroyedEvent` — single `UnitDestroyed` kill detection
- `applyDamageThresholdPSR` — 20+-damage piloting skill roll queueing

Also lifted the seven-member `UnitDestroyed` cause union into an exported
`DestructionCause` type alias shared by the phase and its helpers.

Behavior-preserving: the public `resolveWeaponHit` export keeps the same
name and signature; event ordering and state transitions are unchanged.
Main file drops 548 -> 317 lines. Verified: tsc --noEmit clean, oxlint
0 warnings (max-lines warning resolved), 103/103 simulation/runner tests
pass.
